+++ src/drc/api/tests/test_validation.py
@@ -1,18 +1,26 @@
+import uuid
+from base64 import b64encode
 from copy import deepcopy
+from unittest.mock import patch
 
 from django.test import override_settings
 
+from privates.test import temp_private_root
 from rest_framework import status
 from rest_framework.test import APITestCase
-from vng_api_common.tests import JWTAuthMixin, get_validation_errors
+from vng_api_common.tests import JWTAuthMixin, get_validation_errors, reverse
 from vng_api_common.validators import URLValidator
+from zds_client.tests.mocks import mock_client
 
 from drc.datamodel.constants import OndertekeningSoorten, Statussen
+from drc.datamodel.models import ObjectInformatieObject
 from drc.datamodel.tests.factories import EnkelvoudigInformatieObjectFactory
 
-from .utils import reverse, reverse_lazy
+from .utils import reverse_lazy
 
 INFORMATIEOBJECTTYPE = 'https://example.com/informatieobjecttype/foo'
+ZAAK = 'https://zrc.nl/api/v1/zaken/1234'
+BESLUIT = 'https://brc.nl/api/v1/besluiten/4321'
 
 
 class EnkelvoudigInformatieObjectTests(JWTAuthMixin, APITestCase):
@@ -39,7 +47,7 @@ class EnkelvoudigInformatieObjectTests(JWTAuthMixin, APITestCase):
                 self.assertEqual(error['code'], code)
 
     @override_settings(LINK_FETCHER='vng_api_common.mocks.link_fetcher_404')
-    def test_validate_informatieobjecttype_invalid(self):
+    def test_validate_informatieobjecttype_invalid_url(self):
         url = reverse('enkelvoudiginformatieobject-list')
 
         response = self.client.post(url, {
@@ -49,6 +57,25 @@ class EnkelvoudigInformatieObjectTests(JWTAuthMixin, APITestCase):
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         error = get_validation_errors(response, 'informatieobjecttype')
         self.assertEqual(error['code'], URLValidator.code)
+
+    @override_settings(LINK_FETCHER='vng_api_common.mocks.link_fetcher_200')
+    def test_validate_informatieobjecttype_invalid_resource(self):
+        responses = {
+            INFORMATIEOBJECTTYPE: {
+                'some': 'incorrect property'
+            }
+        }
+
+        url = reverse('enkelvoudiginformatieobject-list')
+
+        with mock_client(responses):
+            response = self.client.post(url, {
+                'informatieobjecttype': INFORMATIEOBJECTTYPE,
+            })
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        error = get_validation_errors(response, 'informatieobjecttype')
+        self.assertEqual(error['code'], 'invalid-resource')
 
     def test_link_fetcher_cannot_connect(self):
         url = reverse('enkelvoudiginformatieobject-list')
@@ -122,6 +149,65 @@ class EnkelvoudigInformatieObjectTests(JWTAuthMixin, APITestCase):
 
         self.assertGegevensGroepValidation(url, 'ondertekening', base_body, cases)
 
+    @temp_private_root()
+    @override_settings(LINK_FETCHER='vng_api_common.mocks.link_fetcher_200')
+    @patch("vng_api_common.validators.fetcher")
+    @patch("vng_api_common.validators.obj_has_shape", return_value=True)
+    def test_inhoud_incorrect_padding(self, *mocks):
+        url = reverse('enkelvoudiginformatieobject-list')
+        content = {
+            'identificatie': uuid.uuid4().hex,
+            'bronorganisatie': '159351741',
+            'creatiedatum': '2018-06-27',
+            'titel': 'detailed summary',
+            'auteur': 'test_auteur',
+            'formaat': 'txt',
+            'taal': 'eng',
+            'bestandsnaam': 'dummy.txt',
+            # Remove padding from the base64 data
+            'inhoud': b64encode(b'some file content').decode('utf-8')[:-1],
+            'link': 'http://een.link',
+            'beschrijving': 'test_beschrijving',
+            'informatieobjecttype': INFORMATIEOBJECTTYPE,
+            'vertrouwelijkheidaanduiding': 'openbaar',
+        }
+
+        # Send to the API
+        response = self.client.post(url, content)
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        error = get_validation_errors(response, 'inhoud')
+        self.assertEqual(error['code'], 'incorrect-base64-padding')
+
+    @temp_private_root()
+    @override_settings(LINK_FETCHER='vng_api_common.mocks.link_fetcher_200')
+    @patch("vng_api_common.validators.fetcher")
+    @patch("vng_api_common.validators.obj_has_shape", return_value=True)
+    def test_inhoud_correct_padding(self, *mocks):
+        url = reverse('enkelvoudiginformatieobject-list')
+        content = {
+            'identificatie': uuid.uuid4().hex,
+            'bronorganisatie': '159351741',
+            'creatiedatum': '2018-06-27',
+            'titel': 'detailed summary',
+            'auteur': 'test_auteur',
+            'formaat': 'txt',
+            'taal': 'eng',
+            'bestandsnaam': 'dummy.txt',
+            # Remove padding from the base64 data
+            'inhoud': b64encode(b'some file content').decode('utf-8'),
+            'link': 'http://een.link',
+            'beschrijving': 'test_beschrijving',
+            'informatieobjecttype': INFORMATIEOBJECTTYPE,
+            'vertrouwelijkheidaanduiding': 'openbaar',
+        }
+
+        # Send to the API
+        response = self.client.post(url, content)
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
 
 @override_settings(LINK_FETCHER='vng_api_common.mocks.link_fetcher_200')
 class InformatieObjectStatusTests(JWTAuthMixin, APITestCase):
@@ -129,7 +215,9 @@ class InformatieObjectStatusTests(JWTAuthMixin, APITestCase):
     url = reverse_lazy('enkelvoudiginformatieobject-list')
     heeft_alle_autorisaties = True
 
-    def test_ontvangen_informatieobjecten(self):
+    @patch("vng_api_common.validators.fetcher")
+    @patch("vng_api_common.validators.obj_has_shape", return_value=True)
+    def test_ontvangen_informatieobjecten(self, *mocks):
         """
         Assert certain statuses are not allowed for received documents.
 
@@ -220,3 +308,50 @@ class FilterValidationTests(JWTAuthMixin, APITestCase):
             with self.subTest(query_param=key, value=value):
                 response = self.client.get(url, {key: value}, HTTP_ACCEPT_CRS='EPSG:4326')
                 self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+
+@override_settings(
+    LINK_FETCHER='vng_api_common.mocks.link_fetcher_200',
+)
+class ObjectInformatieObjectValidationTests(JWTAuthMixin, APITestCase):
+    heeft_alle_autorisaties = True
+
+    list_url = reverse(ObjectInformatieObject)
+
+    @patch('vng_api_common.validators.obj_has_shape', return_value=False)
+    def test_create_oio_invalid_resource_zaak(self, *mocks):
+
+        eio = EnkelvoudigInformatieObjectFactory.create()
+        eio_url = reverse('enkelvoudiginformatieobject-detail', kwargs={
+            'uuid': eio.uuid
+        })
+
+        response = self.client.post(self.list_url, {
+            'object': ZAAK,
+            'informatieobject': f'http://testserver{eio_url}',
+            'objectType': 'zaak'
+        })
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        error = get_validation_errors(response, 'object')
+        self.assertEqual(error['code'], 'invalid-resource')
+
+    @patch('vng_api_common.validators.obj_has_shape', return_value=False)
+    def test_create_oio_invalid_resource_besluit(self, *mocks):
+
+        eio = EnkelvoudigInformatieObjectFactory.create()
+        eio_url = reverse('enkelvoudiginformatieobject-detail', kwargs={
+            'uuid': eio.uuid
+        })
+
+        response = self.client.post(self.list_url, {
+            'object': BESLUIT,
+            'informatieobject': f'http://testserver{eio_url}',
+            'objectType': 'besluit'
+        })
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        error = get_validation_errors(response, 'object')
+        self.assertEqual(error['code'], 'invalid-resource')